pyproject.toml
[tool.poetry]
name = "votemapper"
version = "0.1.0"
description = "Hell Let Loose votemap improver bot for CRCON."
authors = ["Rich Tebb <rich@tebb.io>"]
readme = "README.md"
packages = [
    { include = "app", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
attrs = "^24.2.0"
websockets = "^14.0"
uvloop = "^0.21.0"
python-dotenv = "^1.0.1"
cattrs = "^24.1.2"
aiohttp = "^3.10.10"
yarl = "^1.18.0"
async-cache = "^1.1.1"
numpy = "^2.2.0"
pandas = "^2.2.3"
environ-config = "^24.1.0"
lagom = "^2.7.5"
typeguard = "^4.4.1"
cachetools = "^5.5.0"
wrapt = "^1.17.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
mypy = "^1.13.0"
pyright = "^1.1.388"
pytest = "^8.3.3"
aioresponses = "^0.7.7"
pytest-asyncio = "^0.24.0"
pytest-describe = "^2.2.0"
ipython = "^8.30.0"
ipykernel = "^6.29.5"
pandas-stubs = "^2.2.3.241126"
types-cachetools = "^5.5.0.20240820"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
myapp = "app.__main__:main"

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.pyright]
include = ["src"]
exclude = ["scripts", "docs", "tests", "out"]
reportMissingImports = "error"
reportOptionalMemberAccess = false
reportGeneralTypeIssues = true
reportOptionalSubscript = false
reportUnboundVariable = true
pythonVersion = "3.12"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["scripts", "docs", "tests", "out", "play"]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"

[tool.ruff]
exclude = ["scripts", "docs", "out", "play"]
cache-dir = ".ruff_cache"
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "W", "F", "T"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
known-local-folder = ["src"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402"]
